Add configurable ADC trigger frequency
Adjust the ADC trigger frequency to 1/4 of the master clock, which is configurable from
the user application.

Earlier this trigger frequency was set to 1/4 of the CFG_FM_FREQUENCY_HZ.
This can cause an unmatched frequency if the master clock is set to something else than
CFG_FM_FREQUENCY_HZ, defined in tcd1304_conf.h. To sanity check in later commits.