Use One-Pulse-Timer functionality to avoid timing problems
Background:
The ICG timer generates an ICG pulse. When this timer overflows, an interrupt is generated.
In this interrupt handler, the ADC Trigger timer is started. Remember that the DMA is connected to the ADC in
CIRCULAR mode. When the DMA has transfered 3694 ADC samples to RAM, a new interrupt is generated.
In this interrupt handler the ADC trigger timer MUST be disabled to avoid false restart of ADC sampling.

The problem:
The ADC runs up to 1 MSamples/s with 4 MHz master clock. This means that when the last sample of the
ADC is transfered by the DMA, we have 1 us to disable the ADC trigger timer to stop the DMA.
The STM32F7 running at 216 MHz from FLASH needs approximately 100 ns to start execution in interrupt
handlers.
In an MCU system it can possibly be many interrupts enabled and with different interrupt priority levels.
Also, the interrupts can be temporary disabled for a short time to handle a critical section in the software.
This can lead to the situation where the DMA has restarted a new transfer to RAM. How far the DMA
has moved the data is dependent on how long time the interrupt handler is delayed.
This causes misaligned spectrum data.

Solution:
The STM32 timers has a function called One-Pulse-Timer. This timer is programmed to generate
3694 pulses when triggered. We use software trigger, by enabling the CEN bit in timer CR1 register.
When this timer has generated 3694 pulses, it automatically shuts down the counter by disabling
the CEN bit in its own CR1 register.
This eliminates the above problem, and the interrupt response time is no longer an issue.
This commit utilize this nice hardware function to avoid any potential issues with data alignment.